# openapi.yaml
openapi: 3.0.3
info:
  title: 唐宋电竞陪玩报单平台API
  description: 唐宋电竞陪玩报单平台后端API文档
  version: 1.0.0
  contact:
    name: 唐宋电竞技术团队
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.tangsong-esports.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.tangsong-esports.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "操作成功"
        data:
          type: object
    PageResponse:
      type: object
      properties:
        list:
          type: array
          items:
            type: object
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10

    # 登录相关
...
    Customer:
      type: object
      properties:
        customer_id:
          type: integer
          example: 1
        account:
          type: string
          example: "customer001"
        customer_name:
          type: string
          example: "测试客户"
        contact_method:
          type: string
          example: "微信: test_wx"
        phone_number:
          type: string
          example: "13900139000"
        member_birthday:
          type: string
          format: date-time
          nullable: true
          example: "1990-01-01T00:00:00Z"
        room_code:
          type: string
          example: "ROOM001"
        additional_info1:
          type: string
        additional_info2:
          type: string
        additional_info3:
          type: string
        notes:
          type: string
          example: "VIP客户"
        status:
          type: string
          enum: ["正常", "禁用", "过期"]
          example: "正常"
        is_email_verified:
          type: boolean
          example: false
        is_phone_verified:
          type: boolean
          example: false
        last_login_at:
          type: string
          format: date-time
          nullable: true
        last_login_ip:
          type: string
          example: "127.0.0.1"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        financial_info:
          $ref: '#/components/schemas/CustomerFinancialInfo'
        preferences:
          $ref: '#/components/schemas/CustomerPreferences'
    CustomerFinancialInfo:
      type: object
      properties:
        customer_id:
          type: integer
        initial_real_charge:
          type: number
          format: decimal
          example: 1000.00
        total_consumption:
          type: number
          format: decimal
          example: 500.00
        total_real_charge:
          type: number
          format: decimal
          example: 1500.00
        current_balance:
          type: number
          format: decimal
          example: 1000.00
    CustomerPreferences:
      type: object
      properties:
        customer_id:
          type: integer
        exclusive_discount_type:
          type: string
          example: "固定折扣"
        platform_boss:
          type: string
          example: "平台老板A"
        exclusive_cs:
          type: string
          example: "客服小王"
    CustomerCreateRequest:
      type: object
      required:
        - account
        - customer_name
      properties:
        account:
          type: string
          example: "customer001"
        password:
          type: string
          minLength: 6
          example: "123456"
          description: "可选，如果不提供则账户无法登录，需要后续设置"
        customer_name:
          type: string
          example: "测试客户"
        contact_method:
          type: string
          example: "微信: test_wx"
        phone_number:
          type: string
          example: "13900139000"
        member_birthday:
          type: string
          format: date-time
          nullable: true
          example: "1990-01-01T00:00:00Z"
        room_code:
          type: string
          example: "ROOM001"
        additional_info1:
          type: string
        additional_info2:
          type: string
        additional_info3:
          type: string
        notes:
          type: string
          example: "VIP客户"
        initial_real_charge:
          type: number
          format: decimal
          example: 1000.00
        exclusive_discount_type:
          type: string
          example: "固定折扣"
        platform_boss:
          type: string
          example: "平台老板A"
        exclusive_cs:
          type: string
          example: "客服小王"
    CustomerRechargeRequest:
      type: object
      required:
        - real_charge_amount
        - payment_method
      properties:
        real_charge_amount:
          type: number
          format: decimal
          example: 500.00
        gift_amount:
          type: number
          format: decimal
          example: 50.00
        payment_method:
          type: string
          enum: ["微信", "支付宝", "银行转账", "平台", "内部", "其他"]
          example: "微信"
        transaction_id:
          type: string
          example: "WX20250715001"
        notes:
          type: string
          example: "充值测试"

    # 订单类别相关
...
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"

    # 客户注册登录相关
...
    AdminResetCustomerPasswordRequest:
      type: object
      required:
        - customer_id
        - new_password
      properties:
        customer_id:
          type: integer
          example: 1
        new_password:
          type: string
          minLength: 6
          example: "newpass123"
paths:
...
  /customers:
    get:
      tags:
        - 客户管理
      summary: 获取客户列表
      description: 分页获取客户列表，支持搜索和筛选
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
        - name: account
          in: query
          description: 账号搜索
          schema:
            type: string
        - name: customer_name
          in: query
          description: 客户名称搜索
          schema:
            type: string
        - name: phone_number
          in: query
          description: 手机号搜索
          schema:
            type: string
        - name: status
          in: query
          description: 状态筛选
          schema:
            type: string
            enum: ["正常", "禁用", "过期"]
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'

    post:
      tags:
        - 客户管理
      summary: 创建客户
      description: 创建新的客户
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'
  /customers/{id}:
    get:
      tags:
        - 客户管理
      summary: 获取客户详情
      description: 根据ID获取客户详细信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 客户ID
          schema:
            type: integer
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'

    put:
      tags:
        - 客户管理
      summary: 更新客户
      description: 更新客户信息
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 客户ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Customer'

    delete:
      tags:
        - 客户管理
      summary: 删除客户
      description: 软删除客户
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 客户ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /customers/{id}/recharge:
    post:
      tags:
        - 客户管理
      summary: 客户充值
      description: 为客户账户充值
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 客户ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRechargeRequest'
      responses:
        '200':
          description: 充值成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /customers/{id}/recharge-history:
    get:
      tags:
        - 客户管理
      summary: 获取客户充值记录
      description: 获取指定客户的充值记录
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 客户ID
          schema:
            type: integer
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResponse'
...
  /customers/reset-password:
    post:
      tags:
        - 客户管理
      summary: 管理员重置客户密码
      description: 管理员可以直接为任何客户重置密码
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminResetCustomerPasswordRequest'
      responses:
        '200':
          description: 密码重置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "请求参数错误"
        '404':
          description: 客户不存在
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "客户不存在"
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "密码重置失败"